Cache Works With Singleton
Screen goes back and forth 13 times before OOM.
